model/blip2_server/requirements.txt
fastapi
uvicorn
torch
transformers
Pillow

